Move renderer.js and DevToolsFiberComponentStack.js into backend/fiber
These are at the same level as backend/legacy.

This clarifies that anything outside of this folder shouldn't ideally refer
to a "Fiber".
